requirements.txt
cachetools==2.0.1
certifi==2017.7.27.1
cffi==1.14.6
chardet==3.0.4
click==8.0.1
Flask==2.0.1
google-api-core == 2.7.1
google-auth==2.6.2
google-cloud-core == 2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
gunicorn==19.7.1
idna==2.6
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.3.1
protobuf==3.13
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyparsing==2.4.7
pysha3==1.0.2
pytz==2017.2
requests==2.18.4
rsa==3.4.2
six==1.13
toolz==0.11.1
urllib3==1.22
Werkzeug==2.0.1
